Split State interface declaration from root reducer
